#36 검색기능 키워드 삭제 키워드를 못가져와서 검색이 안됨

+++ src/main/java/com/portfolio/ohousev1/service/post/PostCommentService.java
@@ -14,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -51,7 +52,10 @@ public class PostCommentService {
     }
     @Transactional
     public void deletePostComment(Long postCommentId, String email){
-        postCommentRepository.deleteByIdAndMember_Email(postCommentId, email);
+        Optional<PostComment> comment = postCommentRepository.findById(postCommentId);
+        if (comment.isPresent()) {
+            postCommentRepository.deleteByIdAndMember_Email(postCommentId, email);
+        }
     }
